stock_market_api/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
django-cors-headers>=4.3.0
apscheduler>=3.10.0
openai>=1.3.0
tensorflow>=2.13.0
scikit-learn>=1.3.0